TearBridge test + testnet deploys
