+++ docs/elin/contracts-4.5.0/utils/Strings.md
@@ -1,0 +1,12 @@
+# Strings
+
+
+
+
+
+
+
+*String operations.*
+
+
+
