+++ docs/messaging/interfaces/ISynMessagingReceiver.md
@@ -13,7 +13,7 @@
 ### executeMessage
 
 ```solidity
-function executeMessage(bytes32 _srcAddress, uint256 _srcChainId, bytes _message, address _executor) external nonpayable returns (enum ISynMessagingReceiver.MsgExecutionStatus)
+function executeMessage(bytes32 _srcAddress, uint256 _srcChainId, bytes _message, address _executor) external nonpayable
 ```
 
 Called by MessageBus 
@@ -28,12 +28,6 @@ Called by MessageBus
 | _srcChainId | uint256 | The source chain ID where the transfer is originated from |
 | _message | bytes | Arbitrary message bytes originated from and encoded by the source app contract |
 | _executor | address | Address who called the MessageBus execution function |
-
-#### Returns
-
-| Name | Type | Description |
-|---|---|---|
-| _0 | enum ISynMessagingReceiver.MsgExecutionStatus | undefined |
